LU-4856 misc: Reduce exposure to overflow on page counters.
When the number of an object in use or circulation is tied to memory
size of the system, very large memory systems can overflow 32 bit
counters.  This patch addresses overflow on page counters in the osc
LRU and obd accounting.

Lustre-change: http://review.whamcloud.com/10537
Lustre-commit: e12bbee8207c41b2be5c8bb848f03bb81e75b025

+++ lustre/obdclass/linux/linux-sysctl.c
@@ -231,7 +231,7 @@ int LL_PROC_PROTO(proc_max_dirty_pages_in_mb)
 		/* Don't allow them to let dirty pages exceed 90% of system
 		 * memory and set a hard minimum of 4MB. */
 		if (obd_max_dirty_pages > ((totalram_pages / 10) * 9)) {
-			CERROR("Refusing to set max dirty pages to %u, which "
+			CERROR("Refusing to set max dirty pages to %lu, which "
 			       "is more than 90%% of available RAM; setting "
 			       "to %lu\n", obd_max_dirty_pages,
 			       ((totalram_pages / 10) * 9));
@@ -399,14 +399,14 @@ static struct ctl_table obd_table[] = {
                 .proc_handler = &proc_alloc_fail_rate
         },
 #endif
-        {
-                INIT_CTL_NAME(OBD_MAX_DIRTY_PAGES)
-                .procname = "max_dirty_mb",
-                .data     = &obd_max_dirty_pages,
-                .maxlen   = sizeof(int),
-                .mode     = 0644,
-                .proc_handler = &proc_max_dirty_pages_in_mb
-        },
+	{
+		INIT_CTL_NAME(OBD_MAX_DIRTY_PAGES)
+		.procname = "max_dirty_mb",
+		.data     = &obd_max_dirty_pages,
+		.maxlen   = sizeof(unsigned long),
+		.mode     = 0644,
+		.proc_handler = &proc_max_dirty_pages_in_mb
+	},
 	{
 		INIT_CTL_NAME(OBD_BULK_TIMEOUT)
 		.procname	= "bulk_timeout",
